Compila

+++ BLE/EmisoraBLE.h
@@ -36,9 +36,9 @@ class EmisoraBLE{
    Bluefruit.ScanResponse.addName();
     }
 
-   void anunciarCO(uint8_t temperatura, uint8_t humedad, uint8_t medida){
+   void anunciarCO(int temperatura, int humedad, int medida){
 
-  uint16_t elMajor=((uint16_t)temperatura << 8) | humedad;
+  uint16_t elMajor=((uint16_t)temperatura << 8) | (uint8_t)humedad;
   uint16_t elMinor=(uint16_t)medida;
